requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
beautifulsoup4==4.12.2
pydantic==2.5.0
python-dotenv==1.0.0
google-generativeai==0.3.1
celery==5.3.4
redis==5.0.1
tenacity==8.2.3
pytz==2023.3
structlog==23.2.0
pymongo==4.6.0
